backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
psycopg>=3.2
python-dotenv==1.0.1
pydantic>=2.10
requests==2.32.3